chatGPT/requirements.txt
requests==2.31.0
tiktoken==0.4.0
PyPDF2==3.0.1
python-docx==0.8.11
openai==0.28.0
langchain==0.0.285
llama-index==0.8.23.post1
mysql-connector-python==8.1.0
Flask==2.3.3
elasticsearch==8.9.0
elasticsearch_dsl==8.9.0
flask_cors==4.0.0